补充版权信息，增加mock日志
