+++ Blog/Mango.Service.Blog/Repositories/ArticleDetailRepository.cs
@@ -1,4 +1,4 @@
-﻿/*--------------------------------------------------------------------------*/
+﻿/*--------------------------------------------------------------------------
 //
 //  Copyright 2020 Chiva Chen
 //
