+++ Test/Test/Blog/ArticleService/BlogArticleServiceTest.cs
@@ -1,4 +1,22 @@
-﻿using Mango.Core.Enums;
+﻿/*--------------------------------------------------------------------------
+//
+//  Copyright 2020 Chiva Chen
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+/*--------------------------------------------------------------------------*/
+
+using Mango.Core.Enums;
 using Mango.EntityFramework;
 using Mango.EntityFramework.Abstractions;
 using Mango.Service.Blog.Abstractions.Models.Dto;
@@ -36,26 +54,7 @@ namespace Test.Blog.ArticleService
 
         public override void InitMock(BlogDbContext dbContext, IServiceCollection services)
         {
-            var log = new Mock<ILogger<Mango.Service.Blog.Services.ArticleService>>();
-            log.Setup(x => x.Log(
-                    It.IsAny<LogLevel>(),
-                    It.IsAny<EventId>(),
-                    It.IsAny<It.IsAnyType>(),
-                    It.IsAny<Exception>(),
-                    (Func<It.IsAnyType, Exception, string>)It.IsAny<object>()))
-                    .Callback(new InvocationAction(invocation =>
-                    {
-                        var logLevel = (LogLevel)invocation.Arguments[0]; // The first two will always be whatever is specified in the setup above
-                        var eventId = (EventId)invocation.Arguments[1];  // so I'm not sure you would ever want to actually use them
-                        var state = invocation.Arguments[2];
-                        var exception = (Exception?)invocation.Arguments[3];
-                        var formatter = invocation.Arguments[4];
-
-                        var invokeMethod = formatter.GetType().GetMethod("Invoke");
-                        var logMessage = (string?)invokeMethod?.Invoke(formatter, new[] { state, exception });
-
-                        _output.WriteLine(logMessage);
-                    }));
+            var log = GetMockLogger<Mango.Service.Blog.Services.ArticleService>();
             var articleCache = new Mock<IArticleCacheService>();
             var work = new EfContextWork<BlogDbContext>(dbContext);
             var articleRepository = new ArticleRepository(dbContext);
